+++ src/nn/layers/input_layer.rs
@@ -9,10 +9,10 @@ pub struct InputLayer {
 impl InputLayer {
     /// Creates input layer based on inputs supplied.
     /// Allows for automatic shaping of succeeding layer generation.
-    pub fn from_vec(values: Vec<Vec<f64>>) -> InputLayer {
+    pub fn from_vec(values: Vec<Vec<f32>>) -> InputLayer {
         let rows = values.len();                                                // Number of inputs
         let columns = values.len() / rows;                                      // Number of features.
-        let values: Vec<f64> = values.into_iter().flat_map(|v| v).collect();
+        let values: Vec<f32> = values.into_iter().flat_map(|v| v).collect();
 
         InputLayer {
             input_matrix: Matrix::from(rows, columns, values)
